requirements.txt
Django==3.1.5
gunicorn==20.0.4
psycopg2-binary==2.8.4
djangorestframework==3.12.2
django-environ==0.4.5
django-cors-headers==3.6.0
django-filter==2.4.0
djangorestframework-simplejwt==4.6.0
django-phone-field==1.8.1